Communicate UDP address details
Both the server and the client should send each other net.udp
messages with their respective UDP socket port numbers as the
message entity.

The server will attempt to allocate a "channel" on the UDP socket
for the client which is used for filtering incoming packets. The
channel is associated with the Client so that the server knows where
to dispatch UDP packets later.