Centralise all SDL event handling
This adds the handleEvents method to the client which will process all
available. Previously only one event would be handled per-frame which is
pretty poor as it meant responding to QUIT events could take a long time.

This still needs some work as there's way, way too much going on inside
handleEvents -- as there was in input() -- but that's a task for another day.