"Fix" :see_no_evil: segfault when processing messages
I believe py_entity is somehow being double-freed or something to that
effect. But I'm not entire sure, and I find where that would be happening
even if it was. So for now I'm just gonna pretend py_entity doesn't need
no DECREF and accept the possibility of a memory leak in preference to a
guaranteed segfault.