Handle entity.state messages
This is extremely lazy. Basically it just converts the EntityState struct
to a json11 JSON object and then uses the already existing method on the
entity collection to deal with it.

In future this should happen. Instead the entity collection should be updated
to deal with EntityState objects instead of JSON ones but that would be a
fairly big change, likely also needing changes in the server.